requirements.txt
Flask==3.0.3
gunicorn==22.0.0
Pillow==10.4.0
qrcode==7.4.2